// src/app/model/member.model.ts
export class MemberResponse {
  id: string = "";
  fullName: string = "";
  shortName: string = "";
  startDate: string = "";
  endDate: string = "";
  ined: boolean = false;
}

export class MemberRequest {
  id: string = "";
  fullName: string = "";
  shortName: string = "";
  startDate: string = "";
  endDate: string = "";
  ined: boolean = false;
}