// src/app/model/committee-member.model.ts
export class CommitteeMemberRequest {
  id: string = "";
  memberId: string = "";
  memberName: string = "";
  startDate: string = "";
  endDate: string = "";
  chair: string = "";
  gsmElectionMinutesNumber: string = "";
}

export class CommitteeMemberResponse {
  id: string = "";
  memberId: string = "";
  memberName: string = "";
  startDate: string = "";
  endDate: string = "";
  chair: string = "";
  gsmElectionMinutesNumber: string = "";
}